lab1/requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
psycopg[binary,pool]==3.2.9
faker==37.3.0
markdown==3.8
gunicorn==23.0.0 